requirements.txt
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
Django==5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-telegram-bot==22.5
sqlparse==0.5.5
typing_extensions==4.15.0
